+++ lib/pah/partials/_heroku.rb
@@ -108,5 +108,19 @@ if would_you_like? "Create Heroku apps?".red
   production_app = HerokuApp.new(@app_name.gsub('_',''), "app")
   staging_app = HerokuApp.new("#{production_app.name}-staging", "staging app", true) if config['staging']
 
+  if config['staging'] && staging_app.name.present?
+    append_to_file '.env', "STAGING_APP: #{staging_app.name}\n"
+    append_to_file '.env', "PRODUCTION_APP: #{production_app.name}\n"
+
+    git add: '.env'
+    git commit: "-qm 'Adding STAGING_APP and PRODUCTION_APP configs.'"
+  else
+    # The STAGING_APP is the default integration destination
+    append_to_file '.env', "STAGING_APP: #{production_app.name}\n"
+
+    git add: '.env'
+    git commit: "-qm 'Adding STAGING_APP config.'"
+  end
+
   production_app.open if config['deploy']
 end